% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsheet.R
\name{appsheet}
\alias{appsheet}
\title{Appsheet Function}
\usage{
appsheet(
  tableName,
  Action = "Find",
  Properties = list(Locale = "en-US"),
  Rows = list(),
  Selector = NULL,
  appId = Sys.getenv("APPSHEET_APP_ID"),
  access_key = Sys.getenv("APPSHEET_APP_ACCESS_KEY")
)
}
\arguments{
\item{tableName}{The name of the table to perform actions on.}

\item{Action}{The action to be performed on the table. Default is "Find".}

\item{Properties}{A list of properties for the action. Default is a list with Locale set to "en-US".}

\item{Rows}{A list of rows for the action. Default is an empty list.}

\item{appId}{The AppSheet application ID. Default is retrieved from the APPSHEET_APP_ID environment variable.}

\item{access_key}{The AppSheet application access key. Default is retrieved from the APPSHEET_APP_ACCESS_KEY environment variable.}
}
\value{
A data frame containing the response from the AppSheet API.
}
\description{
This function interacts with the AppSheet API to perform actions on a specified table.
}
\examples{
app_data <- appsheet("my_table", Action = "Find", Properties = list(Locale = "en-US"))

}
